Allow overriding Cross-Origin Resource Sharing header
This makes it possible to use a different URL for connecting to the API
and still allowing connections to the image recognition server from the
pro client.

Ideally I'd have made this configurable, but I couldn't find a way to do
that. res.header 'Access-Control-Allow-Origin' doesn't allow setting
multiple domains in it.